events/requirements.txt
six
Flask==2.1.2
Werkzeug==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
requests==2.28.0
SQLAlchemy==1.4.37
PyJWT==2.4.0
marshmallow-sqlalchemy==0.28.0
Flask-Babel==2.0.0
python-dotenv